% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{assign_kinases_guided}
\alias{assign_kinases_guided}
\title{Assign kinases, given a network and a chip type and additional essential kinases}
\usage{
assign_kinases_guided(
  network,
  chiptype,
  identifier = "Gene_Symbol",
  guided = NULL
)
}
\arguments{
\item{network}{A network output from bnlearn}

\item{chiptype}{Either PTK or STK}

\item{identifier}{The identifier to use in outputs. Can be either "Gene_Symbol" or "Kinase"}

\item{guided}{A vector of Kinases or Gene_Symbols that must be included in the network. The vector must be aligned with what was specified in _identifier_}
}
\value{
A dataframe with upstream kinases assigned to the peptide
}
\description{
Assign kinases, given a network and a chip type and additional essential kinases
}
\examples{
TRUE
}
