% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{candidate_kinases}
\alias{candidate_kinases}
\title{Generate Candidate Kinases Based On Links}
\usage{
candidate_kinases(peptide, arcs, assigned_kinases)
}
\arguments{
\item{peptide}{Peptide ID}

\item{arcs}{The arcs in the network}

\item{assigned_kinases}{The candidate kinases}
}
\value{
common kinases
}
\description{
Generate Candidate Kinases Based On Links
}
\examples{
TRUE
}
