% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_kinased_graphs.R
\name{compare_kinased_graphs}
\alias{compare_kinased_graphs}
\title{Compare two kinased graphhs}
\usage{
compare_kinased_graphs(
  reference,
  comparison,
  ref_name = "Reference",
  comp_name = "Comparison",
  render = FALSE
)
}
\arguments{
\item{reference}{The reference network output from assign_kinases()}

\item{comparison}{The comparison network output from assign_kinases()}

\item{ref_name}{Name for the reference network}

\item{comp_name}{Name for the comparison network}

\item{render}{logical. Whether a graph should be rendered or not.}
}
\value{
A graph object
}
\description{
This functions takes two given graphs and compares them to see the changes from one to the other
}
\examples{
TRUE
}
