% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{default_parent}
\alias{default_parent}
\alias{default_children}
\title{Return Parents/Children with a Usable Default}
\usage{
default_parent(x, n)

default_children(x, n)
}
\arguments{
\item{x}{A bnlearn network}

\item{n}{A node}
}
\value{
Parents or children, given the function. NA if no parents or no children
}
\description{
Return Parents/Children with a Usable Default
}
\section{Functions}{
\itemize{
\item \code{default_children}: Return a default if children don't exist
}}

\examples{
TRUE
}
\keyword{internal}
