% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalize_kinase_graphs.R
\name{equalize_kinase_graphs}
\alias{equalize_kinase_graphs}
\title{Equalize the nodesets of two kinased graphs}
\usage{
equalize_kinase_graphs(reference, comparison)
}
\arguments{
\item{reference}{a reference result from assign_kinases()}

\item{comparison}{a comparison result from assign_kinases()}
}
\value{
A list with two elements, reference and comparison, that have the
same nodesets
}
\description{
Equalize the nodesets of two kinased graphs
}
\examples{
TRUE
}
