% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{filter_peptides}
\alias{filter_peptides}
\title{Return a list of significant peptides}
\usage{
filter_peptides(chipdata, threshold = 0.75)
}
\arguments{
\item{chipdata}{an object of class PamchipSTK or PamchipPTK}

\item{threshold}{a lower cutoff for significance}
}
\value{
a list of significant peptides for each class
}
\description{
Return a list of significant peptides
}
\examples{
TRUE
}
