% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peptides.R
\name{fit_standardized}
\alias{fit_standardized}
\title{Do a standardized fit of expression data from a single peptide's activity}
\usage{
fit_standardized(expr)
}
\arguments{
\item{expr}{a dataframe containing the activity data of only one peptide}
}
\value{
a list of three items std_10, std_50 and std_200, of standardized linear coefficients
}
\description{
Do a standardized fit of expression data from a single peptide's activity
}
\examples{
TRUE
}
