% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_probability_matrix.R
\name{generate_probability_matrix}
\alias{generate_probability_matrix}
\title{Generate a Probability Matrix of observing a given peptide/kinase pair}
\usage{
generate_probability_matrix(annotated_df)
}
\arguments{
\item{annotated_df}{tbl_df. A tibble with a list of all peptide/kinase pairs}
}
\value{
A tibble with the unadjusted probabilities
}
\description{
Generate a Probability Matrix of observing a given peptide/kinase pair
}
\examples{
TRUE
}
\keyword{internal}
