% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{generate_triad_links}
\alias{generate_triad_links}
\title{Generate a Triad of Peptides Given A Network}
\usage{
generate_triad_links(arcs, net)
}
\arguments{
\item{arcs}{a dataframe of arcs between peptides}

\item{net}{a bnlearn network}
}
\value{
a dataframe with all combinations of parents and children of a peptide as kinases.
}
\description{
Generate a Triad of Peptides Given A Network
}
\examples{
TRUE
}
\keyword{internal}
