% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{get_intersections}
\alias{get_intersections}
\title{Generate Intersections between kinases}
\usage{
get_intersections(row, column, assigned_kinases)
}
\arguments{
\item{row}{row peptide}

\item{column}{column peptide}

\item{assigned_kinases}{Kinases assigned}
}
\value{
common kinases beween the two groups
}
\description{
Generate Intersections between kinases
}
\examples{
TRUE
}
