% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_data.R
\docType{data}
\name{kinase_interactome}
\alias{kinase_interactome}
\title{Protein-Protein Interactome of humans}
\format{
A dataframe with 137889 rows and 2 columns

\describe{
\item{from}{The Affecting protein}
\item{to}{The Affected Protein}
}
}
\usage{
kinase_interactome
}
\description{
A complete known map of Protein-Protein interactions in the human
interactome
}
\keyword{datasets}
