% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{kinnet_hamming}
\alias{kinnet_hamming}
\title{Calculate the Hamming Distance between two kinased graphs}
\usage{
kinnet_hamming(reference, comparison)
}
\arguments{
\item{reference}{a reference result from assign_kinases()}

\item{comparison}{a comparison result from assign_kinases()}
}
\value{
A positive integer
}
\description{
Calculate the Hamming Distance between two kinased graphs
}
\examples{
TRUE
}
