% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{kinnet_normalized_hamming}
\alias{kinnet_normalized_hamming}
\title{Calculate Normalized Hamming Distance}
\usage{
kinnet_normalized_hamming(reference, comparison)
}
\arguments{
\item{reference}{a reference result from assign_kinases()}

\item{comparison}{a reference result from assign_kinases()}
}
\value{
A positive integer
}
\description{
Calculate Normalized Hamming Distance
}
\examples{
TRUE
}
