% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{kinnet_normalized_shd}
\alias{kinnet_normalized_shd}
\title{Calculate Normalized Structural Hamming Distance}
\usage{
kinnet_normalized_shd(reference, comparison)
}
\arguments{
\item{reference}{a reference result from assign_kinases()}

\item{comparison}{a comparison result from assign_kinases()}
}
\value{
A positive integer
}
\description{
Calculate Normalized Structural Hamming Distance
}
\examples{
TRUE
}
