% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamchip_data.R
\docType{class}
\name{PamchipData-class}
\alias{PamchipData-class}
\title{The Pamchip Data Superclass}
\value{
an object of class PamchipData
}
\description{
The Pamchip Data Superclass
}
\section{Slots}{

\describe{
\item{\code{chip_type}}{character. A string}
}}

\examples{
TRUE
}
