% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamchip_data_stk.R
\docType{class}
\name{PamchipData-STK-class}
\alias{PamchipData-STK-class}
\title{A representation of the PamChip STK Data}
\value{
An object of class PamchipData-STK
}
\description{
A representation of the PamChip STK Data
}
\section{Slots}{

\describe{
\item{\code{BioNavigatorVersion}}{character. A string indicating the version of BioNavigator that generated the dataset}

\item{\code{ImageAnalysisDate}}{character. The date the analysis was conducted on.}

\item{\code{PamGridVersion}}{character. PamGrid version the chip was run on.}

\item{\code{QuantitationType}}{character. The kind of qunatitation analysis performed.}

\item{\code{SampleData}}{tbl_df. A tibble with the observed activity on each peptide}

\item{\code{SampleCharacteristics}}{tbl_df. A tibble with the characteristics of each sample}

\item{\code{RefData}}{tbl_df.}

\item{\code{PositiveControlData}}{tbl_df.}

\item{\code{PeptideIDs}}{character.}

\item{\code{ProcessedData}}{tbl_df.}

\item{\code{DataProcessDate}}{character. The date when the data was processed}
}}

\examples{
TRUE
}
