% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamchip_data_ptk.R
\name{PamchipData_PTK}
\alias{PamchipData_PTK}
\title{Process a file into a usable structure}
\usage{
PamchipData_PTK(dataset)
}
\arguments{
\item{dataset}{character. Path to a file that holds the output of BioNavigator}
}
\value{
An object of class PamchipData-PTK
}
\description{
Process a file into a usable structure
}
\examples{
TRUE
}
