% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamchip_data_stk.R
\name{PamchipData_STK}
\alias{PamchipData_STK}
\title{Process a file into a usable structure}
\usage{
PamchipData_STK(dataset)
}
\arguments{
\item{dataset}{character. Path to a file that holds the output of BioNavigator}
}
\value{
An object of class PamchipData-STK
}
\description{
Process a file into a usable structure
}
\examples{
TRUE
}
