% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_kinases.R
\name{process_parent_child_link}
\alias{process_parent_child_link}
\title{Filter list of kinases based on the parents and children}
\usage{
process_parent_child_link(
  parent,
  node,
  child,
  kinase_annotation,
  interactome,
  identifier
)
}
\arguments{
\item{parent}{Parent node for the given triad}

\item{node}{The node under consideration}

\item{child}{Child node for the given triad}

\item{kinase_annotation}{A dataframe with the peptide to kinase map}

\item{interactome}{A dataframe with the kinase-kinase interactions}
}
\value{
A dataframe of filtered kinases
}
\description{
Filter list of kinases based on the parents and children
}
\examples{
TRUE
}
\keyword{internal}
