% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_data.R
\docType{data}
\name{ptk_annotation}
\alias{ptk_annotation}
\title{Map of PTK PamChip IDs to PTK}
\format{
A dataframe with 991 rows and 6 columns
\describe{
\item{ID}{The ID of the Peptide on the PTK PamChip}
\item{Group}{The Group categorization for the associated kinase from RegPhos 2.0 Annoatation}
\item{Family}{The Group categorization for the associated kinase from RegPhos 2.0 Annoatation}
\item{Subfamily}{The Subfamily categorization for the associated kinase from RegPhos 2.0 Annoatation. Can be NA}
\item{Kinase}{The Kinase tha can potentially phosphorylate this peptide}
\item{Gene_Symbol}{The HGNC Gene Symbol for the Kinase}
}
}
\usage{
ptk_annotation
}
\description{
A dataset that contains a map of all 196 PTK PamChip Peptides
and the PT Kinases that could phosphorylate them in a human sample
}
\keyword{datasets}
