% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_matrix.R
\docType{data}
\name{ptk_probability_matrix}
\alias{ptk_probability_matrix}
\title{PTK Peptide to Kinase Probability Matrix}
\format{
A dataframe with 991 rows and 3 columns

\describe{
\item{peptide}{The peptide ID}
\item{kinase}{Gene Symbol of the kinase}
\item{posterior}{Posterior probability of the pair}
}
}
\usage{
ptk_probability_matrix
}
\description{
A dataaframe with the bayesian posteriors of finding a given peptide/kinase pair
}
\keyword{datasets}
