% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_reduced_kinased_graph.R
\name{render_reduced_kinased_graph}
\alias{render_reduced_kinased_graph}
\title{Generate a reduced kianse graph}
\usage{
render_reduced_kinased_graph(analysis_result, title)
}
\arguments{
\item{analysis_result}{Object output from assign_kinases()}

\item{title}{Title of the graph}
}
\value{
a graph object
}
\description{
Generate a reduced kianse graph
}
\examples{
TRUE
}
