% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_kinased_graphs.R
\name{safely_add_node}
\alias{safely_add_node}
\title{Add a node safely to a bn object}
\usage{
safely_add_node(net, node)
}
\arguments{
\item{net}{A bn object}

\item{node}{A node label}
}
\value{
A bn object with the node added
}
\description{
This function takes a bn object and a node and adds it if it's not already in the nodeset.
}
\examples{
TRUE
}
