% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_peptides.R
\docType{data}
\name{selected_peptides}
\alias{selected_peptides}
\title{List of significantly differentially expressed peptides in the dataset}
\format{
A dataframe with 833 rows and 4 columns

\describe{
\item{cell_line}{The cell line for the experiment}
\item{reference}{The reference sample}
\item{comparison}{The comparison sample}
\item{peptides}{The peptides selected for that pairing}
}
}
\usage{
selected_peptides
}
\description{
A list of all the peptides found to be differentially phosphorylated
in each pair of samples
}
\keyword{datasets}
