% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_methods.R
\name{stk_accessors}
\alias{stk_accessors}
\alias{pheno_data}
\alias{pheno_data,PamchipData-STK-method}
\alias{exp_data}
\alias{exp_data,PamchipData-STK-method}
\alias{classes}
\alias{classes,PamchipData-STK-method}
\alias{peptides}
\alias{peptides,PamchipData-STK-method}
\title{STK Data Accessor Functions}
\usage{
pheno_data(chipdata, ...)

\S4method{pheno_data}{`PamchipData-STK`}(chipdata)

exp_data(chipdata, ...)

\S4method{exp_data}{`PamchipData-STK`}(chipdata)

classes(chipdata, ...)

\S4method{classes}{`PamchipData-STK`}(chipdata)

peptides(chipdata, ...)

\S4method{peptides}{`PamchipData-STK`}(chipdata)
}
\arguments{
\item{chipdata}{an object of class PamchipData-STK or PamchipData-PTK}

\item{...}{Currently unused}
}
\value{
The requested object
}
\description{
These functions provide a variety of data setters and getter for the Pamchip objects.
}
\details{
These functions allow you to get and set the slots of the object.

\itemize{
\item processed_data accesses the processed and transformed data
\item pheno_data gives the sample characteristics
\item exp_data access the actual intensity values
\item classes gives the unique classes in the chip
\item peptides gives the reference list of peptides on the chip
}
}
