% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_matrix.R
\docType{data}
\name{stk_probability_matrix_kinase}
\alias{stk_probability_matrix_kinase}
\title{STK Peptide to Kinase Probability Matrix for Kinase Identifiers}
\format{
A dataframe with 10428 rows and 3 columns

\describe{
\item{peptide}{The peptide ID}
\item{kinase}{Gene Symbol of the kinase}
\item{posterior}{Posterior probability of the pair}
}
}
\usage{
stk_probability_matrix_kinase
}
\description{
A dataframe with the Bayesian posteriors of finding a given peptide/kinase pair
}
\keyword{datasets}
