% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data}
\alias{subset_data}
\title{Subset The Kinome Data}
\usage{
subset_data(chipdata, peptides, class)
}
\arguments{
\item{chipdata}{Pamchip-PTK. A Pamchip-PTK Object with all the data available}

\item{peptides}{character. A vector of peptides of interest}

\item{class}{character. A specification of the sample class you want to filter to}
}
\value{
A tibble with the filtered and asinh transformed data
}
\description{
Subset the Kinome data to a given list of peptides and conditions
}
\examples{
TRUE
}
