% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_probability_matrix.R
\name{update_probability_matrix}
\alias{update_probability_matrix}
\title{Calculated Updated Probabilities of each peptide/kinase pair}
\usage{
update_probability_matrix(chiptype, assignment_df)
}
\arguments{
\item{chiptype}{character. Either "STK" or "PTK"}

\item{assignment_df}{tbl_df. A tibble with the potential kianse assignments}
}
\value{
an updated dataframe with the probabilities
}
\description{
Calculated Updated Probabilities of each peptide/kinase pair
}
\examples{
TRUE
}
