% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_probability_matrix.R
\name{generate_posterior_probability_df}
\alias{generate_posterior_probability_df}
\title{Generate Posterior probabilities given priors}
\usage{
generate_posterior_probability_df(annotated_df)
}
\arguments{
\item{annotated_df}{tbl_df. A tibble with a list of all peptide/kinase pairs}
}
\value{
A tibble with posterior probabilities for each peptide/kinase pair
}
\description{
Generate Posterior probabilities given priors
}
\examples{
TRUE
}
\keyword{internal}
